// src/models/Acao/AcaoEnum.ts
export enum Cotovelo {
  Repouso = 1,
  LevementeContraido = 2,
  Contraido = 3,
  FortementeContraido = 4,
}

export enum Pulso {
  RotacaoMenos90 = 1,
  RotacaoMenos45 = 2,
  Repouso = 3,
  Rotacao45 = 4,
  Rotacao90 = 5,
  Rotacao135 = 6,
  Rotacao180 = 7,
}

export enum Rotacao {
  RotacaoMenos90 = 1,
  RotacaoMenos45 = 2,
  Repouso = 3,
  Rotacao45 = 4,
  Rotacao90 = 5,
}

export enum Inclinacao {
  ParaCima = 1,
  Repouso = 2,
  ParaBaixo = 3,
}

export enum TipoAcao {
  Cotovelo = 1,
  Pulso = 2,
  Rotacao = 3,
  Inclinacao = 4,
}